feat(project): Adding support for chat on app
