+++ infra/package.json
@@ -31,6 +31,7 @@
     "@packages/cognito": "workspace:*",
     "@packages/models": "workspace:*",
     "@packages/notifications": "workspace:*",
+    "@packages/prompts": "workspace:*",
     "@services/chat": "workspace:*",
     "@services/currencies": "workspace:*",
     "@services/documents": "workspace:*",
