+++ packages/prompts/eslint.config.ts
@@ -1,0 +1,11 @@
+import path from 'path';
+import { fileURLToPath } from 'url';
+
+import { getNodeConfig } from '@packages/config/eslint';
+
+const __dirname = path.dirname(fileURLToPath(import.meta.url));
+
+export default [
+  { ignores: ['node_modules/', 'dist/'] },
+  ...getNodeConfig(__dirname),
+];
