+++ packages/prompts/jest.config.ts
@@ -1,0 +1,16 @@
+import type { Config } from 'jest';
+
+const config: Config = {
+  testEnvironment: 'node',
+  roots: ['<rootDir>/src'],
+  testMatch: ['**/*.test.ts'],
+  transform: {
+    '^.+\\.ts$': ['ts-jest', { useESM: false }],
+  },
+  moduleNameMapper: {
+    '^@packages/prompts/(.*)$': '<rootDir>/src/$1',
+  },
+  moduleFileExtensions: ['ts', 'js', 'json'],
+};
+
+export default config;
