+++ packages/prompts/package.json
@@ -1,0 +1,25 @@
+{
+  "name": "@packages/prompts",
+  "version": "1.0.0",
+  "private": true,
+  "description": "Single source of truth for AI prompt templates, Bedrock model routing and prompt output contracts",
+  "scripts": {
+    "typecheck": "tsc --noEmit",
+    "lint": "eslint .",
+    "lint:fix": "eslint . --fix",
+    "test": "jest"
+  },
+  "keywords": [],
+  "author": "Miguel Angel Gutierrez Maya",
+  "license": "ISC",
+  "packageManager": "pnpm@11.1.1",
+  "devDependencies": {
+    "@packages/config": "workspace:*",
+    "@types/jest": "catalog:",
+    "@types/node": "catalog:",
+    "eslint": "catalog:",
+    "jest": "catalog:",
+    "ts-jest": "catalog:",
+    "typescript": "catalog:"
+  }
+}
