+++ packages/prompts/src/bedrock/models.ts
@@ -1,5 +1,5 @@
 /**
- * Bedrock model identifiers used by the chat orchestration.
+ * Bedrock model identifiers used by the AI orchestration.
  *
  * Two families coexist in Bedrock:
  *  - Amazon Nova (Micro, Lite, Pro) — Amazon's native models, on-demand supported.
@@ -30,3 +30,11 @@ export const BEDROCK_MODELS = {
 
 export type BedrockModelId =
   (typeof BEDROCK_MODELS)[keyof typeof BEDROCK_MODELS];
+
+/**
+ * Foundation-model id behind the Claude cross-region inference profile
+ * (the profile id minus the `us.` region prefix). IAM grants need it because
+ * the profile fans out to the underlying foundation models per region.
+ */
+export const CLAUDE_HAIKU_FOUNDATION_MODEL_ID =
+  'anthropic.claude-haiku-4-5-20251001-v1:0';
