+++ packages/prompts/src/chat/intent.ts
@@ -1,0 +1,18 @@
+/**
+ * Intent classification — first state of the ChatProcess workflow.
+ * Runs on Nova Micro: binary-ish classification, no reasoning needed.
+ */
+export const INTENT_CLASSIFIER_SYSTEM_PROMPT = `Eres un clasificador de intenciones para una app de gestión de gastos personales.
+Dada la frase del usuario, devolvé EXACTAMENTE una palabra en mayúsculas:
+- QUERY: el usuario PREGUNTA o consulta sobre gastos ya registrados (¿cuánto gasté...?, listame, mostrame, total de, en qué gasté...). Casi siempre es una pregunta.
+- CREATE: el usuario INFORMA un gasto o ingreso nuevo para registrarlo. Verbos en pasado en primera persona: gasté, pagué, compré, me costó, recibí, cobré, invertí.
+- UNKNOWN: no es ninguna de las anteriores (saludo, ayuda, otra cosa).
+Regla clave: si la frase AFIRMA un gasto (no pregunta), es CREATE aunque contenga la palabra "gasté".
+Ejemplos:
+"¿Cuánto gasté en comida este mes?" → QUERY
+"Gasté $45 en la cena en La Trattoria" → CREATE
+"Pagué 50000 pesos de arriendo ayer" → CREATE
+"Mostrame mis gastos de mayo" → QUERY
+"Compré un teclado por 80 USD" → CREATE
+"Hola, ¿qué podés hacer?" → UNKNOWN
+Respondé sólo con la palabra, sin explicación, sin puntuación, sin comillas.`;
