+++ packages/prompts/tsconfig.json
@@ -1,0 +1,13 @@
+{
+  "extends": "@packages/config/tsconfig.base.json",
+  "compilerOptions": {
+    "outDir": "dist",
+    "rootDir": ".",
+    "baseUrl": ".",
+    "paths": {
+      "@packages/prompts/*": ["./src/*"]
+    }
+  },
+  "include": ["src", "eslint.config.ts", "jest.config.ts"],
+  "exclude": ["node_modules", "dist"]
+}
