+++ services/chat/tsconfig.json
@@ -16,6 +16,7 @@
     "paths": {
       "@services/chat/*": ["./src/*"],
       "@packages/models/*": ["node_modules/@packages/models/src/*"],
+      "@packages/prompts/*": ["node_modules/@packages/prompts/src/*"],
       "@services/expenses/*": ["node_modules/@services/expenses/src/*"],
       "@services/shared/*": ["node_modules/@services/shared/src/*"]
     }
